+++ slytranslate/tests/Unit/ContentTranslationListFastPathTest.php
@@ -49,7 +49,7 @@ class ContentTranslationListFastPathTest extends TestCase {
 						return $this;
 					}
 
-					public function using_temperature( int $temperature ) {
+					public function using_temperature( float $temperature ) {
 						return $this;
 					}
